Support simple transaction editing
1.5.4

Also switches some uses of newer Node features back to older,
basic JS stuff. I'd like to switch to that at some point, but
since it's a breaking change for users on older Node versions,
it'll have to be a major version release. We could also set up
babel, but that's sort of annoying in its own way....